CMake: Prefer static libraries over shared libraries
This option allows us to give cmake a priority when searching for
libraries. Invert the order of library suffixes to be static first then
shared.

This doesn't fully remove the problems of shared library linking but at
least mitigates the problem that #4301 brought up.

As a test I have the system libxxhash installed on my system and cmake
now prefers the static library from that package instead of the shared
library.

Still more work to go, but this at least removes a bit of the headache.