FEXCore: Use musl libm statically
This removes our dependency on libm (although it is still included with
libstdc++ and libgcc). The only place libm is required is actually the
x87 softfloat where we fallback to 128-bit softfloat functions. These
are the couple of transcendental operations that the Berkely Softfloat
library doesn't handle.

This doesn't quite solve the issue that #4301 brought up, but it's
getting closer.