FEXCore: Statically link libgcc and libstdc++
This PR along with the previous ones gets FEXCore and FEXInterpreter
only linking in libc.

```bash
$ ldd ./Bin/FEXInterpreter
        linux-vdso.so.1 (0x0000ffffa206f000)
        libc.so.6 => /lib/aarch64-linux-gnu/libc.so.6 (0x0000ffffa1460000)
        /lib/ld-linux-aarch64.so.1 (0x0000ffffa2020000)
```

The only remaining issue is if the builder has libfmt, then it is
unlikely to have static versions of those libraries available which can
ruin this.

Looks like we'll need to go back to #3709 and #3677 to make that problem
go away.

Additionally xxhash can be installed on the host, but on Ubuntu at least
they provide a statically library as well, which #4410 at least
resolves.