Various: UBSAN fixes around unaligned accesses
Most of the cases we are just allocating objects unaligned so we need to
mark them as packed. The primary outlier is The JITCodeTail which needs
to ensure its SpinLockFutex is aligned to the natural alignment of
uint32_t. This is guaranteed by us allocating the object just after
aligned code, but we need to change the object to be packed to make
UBSAN happy. Just add some assert checks to ensure behaviour doesn't
break here.

Also reordered a couple things in the ELFCodeLoader at the end to ensure
the random number location is correctly aligned. Didn't mess with the
"top" of the stack which actually does need to follow the correct order.