x86Tables: Generate compile time constant DDD and H0F38 tables
These are the first two tables, and they are the easiest to do.

The DDD table is the easiest one to look at, since it's basically a 1:1
translation from the C source to python.

The H0F38 table needed a little bit of work, since we did runtime
selection for CRC/SHA/AES, and that runtime selection is now done in the
handler for the instruction instead. This lets the table be a compile
time constant just like DDD.